requirements.txt
asgiref==3.8.1 ; python_version >= "3.12" and python_version < "4.0"
django-rest-framework==0.1.0 ; python_version >= "3.12" and python_version < "4.0"
django==5.0.4 ; python_version >= "3.12" and python_version < "4.0"
djangorestframework==3.15.1 ; python_version >= "3.12" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
